+++ card_live_dashboard/callbacks/__init__.py
@@ -2,6 +2,8 @@ from typing import List, Set, Dict
 from datetime import datetime, timedelta
 from dash.dependencies import Input, Output, State
 import dash
+from dash.exceptions import PreventUpdate
+import re
 
 from card_live_dashboard.service.CardLiveDataManager import CardLiveDataManager
 from card_live_dashboard.model.CardLiveData import CardLiveData
@@ -54,6 +56,13 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         return is_open
 
     @app.callback(
+        Output('custom-time-period', 'is_open'),
+        [Input('time-period-items', 'value')]
+    )
+    def toggle_custom_time_period(time_period_items):
+        return time_period_items == 'custom'
+
+    @app.callback(
         [Output('global-sample-count', 'children'),
          Output('global-most-recent', 'children'),
          Output('time-period-items', 'options'),
@@ -76,6 +85,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
          Input('organism-lmat-select', 'value'),
          Input('organism-rgi-kmer-select', 'value'),
          Input('time-period-items', 'value'),
+         Input('date-picker-range', 'start_date'),
+         Input('date-picker-range', 'end_date'),
          Input('timeline-type-select', 'value'),
          Input('timeline-color-select', 'value'),
          Input('totals-type-select', 'value'),
@@ -87,6 +98,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                            amr_gene_families: List[str], resistance_mechanisms: List[str],
                            amr_genes: List[str], organism_lmat: str,
                            organism_rgi_kmer: str, time_dropdown: str,
+                           start_date: str, end_date: str,
                            timeline_type_select: str, timeline_color_select: str,
                            totals_type_select: str, totals_color_select: str,
                            rgi_type_select: str, n_intervals):
@@ -100,6 +112,16 @@ def build_callbacks(app: dash.dash.Dash) -> None:
         :param timeline_color_select: The color selection for the timeline.
         :return: The figures to place in the main figure region of the page.
         """
+        custom_date = None
+        if time_dropdown == 'custom':
+            if start_date is not None and end_date is not None:
+                start_date = datetime.strptime(re.split(r'[T ]', start_date)[0], '%Y-%m-%d')
+                end_date = datetime.strptime(re.split(r'[T ]', end_date)[0], '%Y-%m-%d')
+                custom_date = {
+                    'start': start_date,
+                    'end': end_date,
+                }
+
         data = CardLiveDataManager.get_instance().card_data
         global_samples_count = len(data)
         global_last_updated = f'{data.latest_update(): %b %d, %Y}'
@@ -111,7 +133,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                                      resistance_mechanisms=resistance_mechanisms,
                                      amr_genes=amr_genes,
                                      organism_lmat=organism_lmat,
-                                     organism_rgi_kmer=organism_rgi_kmer)
+                                     organism_rgi_kmer=organism_rgi_kmer,
+                                     custom_date=custom_date)
 
         fig_settings = {
             'timeline': {'type': timeline_type_select, 'color': timeline_color_select},
@@ -129,6 +152,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                 'label': f'Last {value} ({time_subsets[value].samples_count()})',
                 'value': value,
             })
+        time_dropdown_text.append({'label': 'Custom', 'value': 'custom'})
 
         samples_count_string = f'{time_subsets[time_dropdown].samples_count()}/{global_samples_count}'
 
@@ -164,7 +188,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
 def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
                   drug_classes: List[str], amr_gene_families: List[str],
                   resistance_mechanisms: List[str], amr_genes: List[str], organism_lmat: str,
-                  organism_rgi_kmer: str) -> Dict[str, CardLiveData]:
+                  organism_rgi_kmer: str, custom_date: Dict[str, datetime]) -> Dict[str, CardLiveData]:
     time_now = datetime.now()
 
     data = data.select(table='rgi', by='cutoff', type='row', level=rgi_cutoff_select) \
@@ -184,6 +208,11 @@ def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
         '6 months': data.select(table='main', by='time', start=time_now - SIX_MONTHS, end=time_now),
         'year': data.select(table='main', by='time', start=time_now - YEAR, end=time_now),
     }
+
+    if custom_date is not None:
+        time_subsets['custom'] = data.select(table='main', by='time', start=custom_date['start'], end=custom_date['end'])
+    else:
+        time_subsets['custom'] = data
 
     return time_subsets